% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{capm}
\alias{capm}
\title{Capital Asset Pricing Model (CAPM)}
\usage{
capm(r, m, f)
}
\arguments{
\item{r}{Numeric vector of returns.}

\item{m}{Numeric vector of market returns.}

\item{f}{Numeric value of risk-free return.}
}
\value{
A matrix with the CAPM betas.
}
\description{
Capital Asset Pricing Model (CAPM)
}
