% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{chol}
\alias{chol}
\alias{chol_mat}
\title{Cholesky decomposition}
\usage{
chol_mat(x, type)
}
\arguments{
\item{x}{Numeric matrix.}

\item{type}{Character string with the type of decomposition.}
}
\value{
A matrix with the Cholesky decomposition.
}
\description{
Cholesky decomposition
}
