% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\docType{package}
\name{cpp11armadillotest-package}
\alias{cpp11armadillotest}
\alias{cpp11armadillotest-package}
\title{cpp11armadillotest: Testing cpp11armadillo}
\description{
Testing cpp11armadillo. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11armadillo/issues}
}

}
\author{
\strong{Maintainer}: YOUR NAME \email{YOUR@EMAIL.COM} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})

}
\keyword{internal}
