% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{eigen_gen}
\alias{eigen_gen}
\alias{eigen_gen_mat}
\alias{eigen_gen_mat_complex_wrapper}
\alias{eigen_gen_dbl_complex_wrapper}
\alias{eigen_gen_no_wrapper}
\title{Eigenvalues and eigenvectors of a general matrix}
\usage{
eigen_gen_mat(x)

eigen_gen_mat_complex_wrapper(x)

eigen_gen_dbl_complex_wrapper(x)

eigen_gen_no_wrapper(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
A matrix with eigenvalues.

A matrix with eigenvalues.

A vector with eigenvalues.

A vector with eigenvalues.
}
\description{
Eigenvalues and eigenvectors of a general matrix
}
