% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{eigen_sym}
\alias{eigen_sym}
\alias{eigen_sym_mat}
\alias{eigen_sym_dbl}
\title{Eigenvalues and eigenvectors of a symmetric matrix}
\usage{
eigen_sym_mat(x)

eigen_sym_dbl(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
A matrix with the eigenvalues.
}
\description{
Eigenvalues and eigenvectors of a symmetric matrix
}
