% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{ols}
\alias{ols}
\alias{ols_mat}
\alias{ols_dbl}
\title{Ordinary Least Squares (OLS) regression}
\usage{
ols_mat(y, x)

ols_dbl(y, x)
}
\arguments{
\item{y}{Numeric vector of dependent variable.}

\item{x}{Numeric matrix of independent variables.}
}
\value{
A matrix with the estimated coefficients.

A vector with the estimated coefficients.
}
\description{
Ordinary Least Squares (OLS) regression
}
