% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{ols_qr}
\alias{ols_qr}
\alias{ols_qr_mat}
\alias{ols_qr_dbl}
\title{Ordinary Least Squares (OLS) regression using QR decomposition}
\usage{
ols_qr_mat(y, x, econ)

ols_qr_dbl(y, x, econ)
}
\arguments{
\item{y}{Numeric vector of dependent variable.}

\item{x}{Numeric matrix of independent variables.}

\item{econ}{Logical value indicating if the economic QR version should be used.}
}
\value{
A matrix with the estimated coefficients.

A vector with the estimated coefficients.
}
\description{
Ordinary Least Squares (OLS) regression using QR decomposition
}
