% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillotest-package.R
\name{typedef}
\alias{typedef}
\alias{typedef_Col_double}
\alias{typedef_Col_int}
\alias{typedef_uvec}
\alias{typedef_Mat_double}
\alias{typedef_Mat_int}
\alias{typedef_SpMat_double}
\alias{typedef_SpMat_int}
\title{Typedefs}
\usage{
typedef_Col_double(x)

typedef_Col_int(x)

typedef_uvec(x)

typedef_Mat_double(x)

typedef_Mat_int(x)

typedef_SpMat_double(x)

typedef_SpMat_int(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A matrix with the input vector.

A matrix with the input vector.

A matrix with the input vector.

A matrix with the input vector.

A matrix with the input vector.

A matrix with the input vector.

A matrix with the input vector.
}
\description{
Typedefs
}
